requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.1.1
qdrant-client==1.12.1
google-generativeai==0.8.3
groq==0.13.1
tavily-python==0.5.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.19
aiofiles==24.1.0
# ── Multimodal processing ──────────────────────────────
PyMuPDF==1.25.1
pytesseract==0.3.13
Pillow==11.1.0
